Wake on the push itself, from the seam that already delivers webhooks
`wake` could only be run by hand or by a timer, which makes the fastest
path to an agent a polling interval. The node host now runs the same
pass from `postReceive` — the seam that already makes a push deliver —
so a review that lands wakes whatever was watching for it.

The pass moves to `server/Wake.node.ts` so the host and the CLI run one
implementation. A wake that fired differently depending on who asked for
it would be a rule an operator cannot reason about.

Forked rather than awaited: `postReceive` runs inside the push, and a
push must not wait on whatever a rule decides to start. Two pushes
landing together are serialized per repository, with a single re-run
remembered rather than a queue — a walk that has not started yet will
see everything that arrived before it does, so more than one pending run
is more than one walk over the same refs. Without that, both would spawn
for the same event, which is the duplicate the bookmark exists to
prevent.

The walk gets its own repository with `hooksNoop`, so a wake that writes
cannot wake itself.

Off unless asked for, as `serve --wake`. What runs is the operator's own
file beside the repository, so nothing a pusher writes chooses the
command — but a server that will start processes at all should have been
told to, and it says so at startup.

The test drives a real server and a real push, and fails when the switch
is off — which is the only way to know it is the wiring under test and
not the walk.
